requirements.txt
aiogram==3.4.1
openai==1.12.0
python-dotenv==1.0.1
aiofiles==23.2.1
jinja2==3.1.3 